backend/main.py
"""
Real-Time Streaming Voice Assistant
FastAPI entry point
"""
import logging
import os

from fastapi import FastAPI
from fastapi.staticfiles import StaticFiles
from fastapi.responses import FileResponse
from dotenv import load_dotenv

# Load .env FIRST before any other imports that read env vars
load_dotenv()

logging.basicConfig(
    level=logging.DEBUG,
    format="%(levelname)s:%(name)s: %(message)s",
)
# Suppress noisy third-party loggers
logging.getLogger("httpx").setLevel(logging.WARNING)
logging.getLogger("httpcore").setLevel(logging.WARNING)
logging.getLogger("openai").setLevel(logging.WARNING)
logging.getLogger("uvicorn.access").setLevel(logging.WARNING)

from backend.websocket_gateway import router as ws_router

app = FastAPI(title="Voice Assistant")

# Mount WebSocket routes
app.include_router(ws_router)

# Serve frontend static files
app.mount("/static", StaticFiles(directory="frontend"), name="static")

@app.get("/")
async def serve_index():
    return FileResponse("frontend/index.html")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "backend.main:app",
        host=os.getenv("HOST", "0.0.0.0"),
        port=int(os.getenv("PORT", 8000)),
        reload=True,
    )
