requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
openai==1.12.0
elevenlabs==1.1.2
aiohttp==3.9.3
aiofiles==23.2.1
